+++ apps/web/src/middleware.ts
@@ -5,7 +5,7 @@ import {
 import { type NextRequest, NextResponse } from 'next/server';
 
 // Known public routes, add more as needed
-const publicRoutes = ['/', '/login', '/join', '/join/secure', '/recovery'];
+const publicRoutes = ['/', '/join', '/join/secure', '/recovery'];
 const publicDynamicRoutes = ['/join/invite'];
 
 export default async function middleware(req: NextRequest) {
